templates/natural_language_to_SQL/requirements.txt
openai>=1.30.5
requests>=2.32.3
python-dotenv
semantic_kernel==1.28.0
#semantic_kernel==0.4.1.dev0
matplotlib
pydantic==2.7.2
pydantic-settings>=2.5.2
pyarrow
numpy
pandas
pytest
rich
chainlit
pyodbc
